Move getCurrentTab() from discover util to reader util
